Verifier always works with u64, remove u8 dependencies

+++ verifier/src/vadcop_final_verifier.rs
@@ -4488,11 +4488,11 @@ fn verifier_info() -> VerifierInfo {
     }
 }
 
-pub fn verify(proof: &VadcopFinalProof, vk: &[u8]) -> bool {
+pub fn verify(proof: &VadcopFinalProof, vk: &[u64]) -> bool {
     stark_verify::<Poseidon16, 16>(&proof.proof_with_publics(), vk, &verifier_info(), q_verify, query_verify)
 }
 
-pub fn verify_bytes(proof: &[u8], vk: &[u8]) -> bool {
+pub fn verify_bytes(proof: &[u64], vk: &[u64]) -> bool {
     stark_verify::<Poseidon16, 16>(proof, vk, &verifier_info(), q_verify, query_verify)
 }
